% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_mc.R
\name{baseline_mc}
\alias{baseline_mc}
\title{Resampling basic variables}
\usage{
baseline_mc(data = fitR2, montecarlo = 1000)
}
\arguments{
\item{data}{input data}

\item{montecarlo}{number of repeated samples for accept-reject algorithm}
}
\value{
data after resampling
}
\description{
Resampling basic variables
}
