% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cen_and_scale}
\alias{cen_and_scale}
\title{Centering and scaling time variable}
\usage{
cen_and_scale(time)
}
\arguments{
\item{time}{time variable}
}
\value{
dataframe with scaled time variable, mean and standard deviation of scaled time variable
}
\description{
Centering and scaling time variable
}
