% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitg.R
\name{fitg}
\alias{fitg}
\title{Fitting model}
\usage{
fitg(
  data,
  boot = FALSE,
  mreg = "binomial",
  lreg = c("binomial", "gaussian", "gaussian"),
  yreg = "binomial",
  dof = 3
)
}
\arguments{
\item{data}{input data}

\item{boot}{doing bootstraping}

\item{mreg}{regression model for mediator}

\item{lreg}{regression model for time-varying variable}

\item{yreg}{regression model for outcome variable}

\item{dof}{degree of freedom}
}
\value{
fitted model
}
\description{
Fitting model
}
