% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_form.R
\name{g_form}
\alias{g_form}
\title{g-formula}
\usage{
g_form(data, ii = 2, length = 12, am = 1, ay = 0)
}
\arguments{
\item{data}{input data}

\item{ii}{patient id}

\item{length}{time to follow up}

\item{am}{am}

\item{ay}{ay}
}
\value{
dataframe with predict to follow up time
}
\description{
g-formula
}
