% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mytheme}
\alias{mytheme}
\title{Helper to adjust styling of a plot}
\usage{
mytheme(...)
}
\arguments{
\item{...}{arbitrary params}
}
\value{
list of updated aesthetic values
}
\description{
Helper to adjust styling of a plot
}
