% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tvmedg}
\alias{plot.tvmedg}
\title{Plot function}
\usage{
\method{plot}{tvmedg}(x, what = c("cumY", "tvY"), ...)
}
\arguments{
\item{x}{data}

\item{what}{Cummulative Y or time-varying Y or both}

\item{...}{arbitrary params}
}
\value{
plot
}
\description{
Plot function
}
