% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resamp}
\alias{resamp}
\title{Resampling input data}
\usage{
resamp(data, boot = FALSE)
}
\arguments{
\item{data}{input data}

\item{boot}{doing boostrap}
}
\value{
an original dataframe if boot = FALSE, resampled dataframe if boot = TRUE
}
\description{
Resampling input data
}
