% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulation dataset}
\format{
A data frame with 1147 rows and 13 variables:
\describe{
  \item{\code{id}}{integer patient id}
  \item{\code{mm}}{double time variable}
  \item{\code{Ap}}{double exposure variable}
  \item{\code{Mp}}{double mediator variable}
  \item{\code{L1}}{double time-varying variable}
  \item{\code{L2}}{double time-varying variable}
  \item{\code{L3}}{double time-varying variable}
  \item{\code{Yp}}{double outcome variable}
  \item{\code{age}}{double age variable}
  \item{\code{sex}}{integer sex variable}
  \item{\code{ow}}{integer overweight}
  \item{\code{risk}}{integer risk}
  \item{\code{lastid}}{double last time follow up}
}
}
\usage{
sim_data
}
\description{
This is a simulation dataset
}
\details{
DETAILS
}
\keyword{datasets}
