% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tvmedg_theme}
\alias{tvmedg_theme}
\title{Helper to adjust styling of a plot}
\usage{
tvmedg_theme(...)
}
\arguments{
\item{...}{arbitrary params}
}
\value{
list of updated aesthetic values
}
\description{
Helper to adjust styling of a plot
}
