% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_form.R
\name{g_form}
\alias{g_form}
\title{g-formula}
\usage{
g_form(data = fitR2$res_df, model = fitR2, followup = 12, am = 1, ay = 0)
}
\arguments{
\item{data}{input data}

\item{model}{model for g formula}

\item{followup}{length of follow up}

\item{am}{am}

\item{ay}{ay}
}
\value{
dataframe with predict to follow up time
}
\description{
g-formula
}
