% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(
  id,
  basec,
  expo,
  med,
  tvar,
  lag,
  outc,
  time,
  norev = NULL,
  tvar_to_med = FALSE,
  cont_exp = FALSE,
  cont_exp_std = FALSE,
  sp_list = NULL,
  sp_type = NULL,
  sp_df = NULL,
  data
)
}
\arguments{
\item{id}{patient id}

\item{basec}{time-fixed variables}

\item{expo}{exposure variable}

\item{med}{mediator variable}

\item{tvar}{time-varying variable}

\item{lag}{number of lag}

\item{outc}{outcome variable}

\item{time}{time variable}

\item{norev}{non-reversible variable (among expo,med,tvar)}

\item{tvar_to_med}{time-varying varible to mediator}

\item{cont_exp}{continuous exposure}

\item{cont_exp_std}{standardize the continuous exposure}

\item{sp_list}{splines list}

\item{sp_type}{splines type}

\item{sp_df}{spines degree of freedom}

\item{data}{input data}
}
\value{
a data frame with lag of exposure, mediator, time-varying variable
}
\description{
Process data
}
