% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmedg.R
\name{tvmedg}
\alias{tvmedg}
\title{Time-varying mediation analysis using g-computation}
\usage{
tvmedg(
  data,
  basec,
  expo,
  med,
  tvar,
  outc,
  time,
  lag = 2,
  norev = NULL,
  cont_exp = NULL,
  cont_exp_std = F,
  tvar_to_med = F,
  mreg = "binomial",
  lreg = c("binomial", "gaussian", "gaussian"),
  yreg = "binomial",
  sp_list = NULL,
  sp_type = NULL,
  sp_df = NULL,
  followup = 12,
  seed = 0,
  montecarlo = 10,
  boot = FALSE,
  nboot = 1,
  ci = 0.95,
  parallel = TRUE
)
}
\arguments{
\item{data}{input data}

\item{basec}{time-fixed variables}

\item{expo}{exposure variable}

\item{med}{mediator variable}

\item{tvar}{time-varying variable}

\item{outc}{outcome variable}

\item{time}{time variable}

\item{lag}{number of lag}

\item{norev}{non-reversible variable (among expo,med,tvar)}

\item{cont_exp}{continuous exposure}

\item{cont_exp_std}{standardize the continuous exposure}

\item{tvar_to_med}{time-varying varible to mediator}

\item{mreg}{regression model for mediator}

\item{lreg}{regression model for time-varying variable}

\item{yreg}{regression model for outcome variable}

\item{sp_list}{splines list}

\item{sp_type}{splines type}

\item{sp_df}{spines degree of freedom}

\item{followup}{length of follow up}

\item{seed}{set seed}

\item{montecarlo}{number of repeated samples for accept-reject algorithm}

\item{boot}{doing boostrap}

\item{nboot}{bootstraping times}

\item{ci}{confident interval}

\item{parallel}{run parallel}
}
\value{
Q11, Q10, Q00
}
\description{
Time-varying mediation analysis using g-computation
}
